presence_test.go
package whatsmeow

import (
	"testing"
	waBinary "go.mau.fi/whatsmeow/binary"

	"github.com/stretchr/testify/assert"
	"go.mau.fi/whatsmeow/types"
)

// MockClient es una implementación de la interfaz Client para las pruebas.
type MockClient struct {
	// Incluye cualquier campo necesario para las pruebas
}

// MockAttrGetter es una implementación de la interfaz AttrGetter para las pruebas.
type MockAttrGetter struct {
	// Incluye cualquier campo necesario para las pruebas
}

func (m *MockAttrGetter) OK() bool {
	return true
}

func (m *MockAttrGetter) Errors() []error {
	return nil
}

// MockNode es una implementación de la interfaz Node para las pruebas.
type MockNode struct {
	// Incluye cualquier campo necesario para las pruebas
}

func (m *MockNode) AttrGetter() waBinary.AttrGetter {
	return &MockAttrGetter{}
}

func (m *MockNode) GetChildren() []waBinary.Node {
	return nil
}

func (m *MockNode) Tag() string {
	return ""
}

func (m *MockNode) Content() []byte {
	return nil
}

// Utilidad para inicializar un MockClient para las pruebas
func setupMockClient() *MockClient {
	return &MockClient{}
}

func TestSendPresence(t *testing.T) {
	cli := setupMockClient()

	// Caso: len(cli.Store.PushName) == 0
	cli.Store.PushName = ""
	err := cli.SendPresence(types.PresenceAvailable)
	assert.EqualError(t, err, ErrNoPushName.Error())

	// Caso: state == types.PresenceAvailable
	cli.Store.PushName = "John Doe"
	err = cli.SendPresence(types.PresenceAvailable)
	assert.NoError(t, err) // Verifica que no hay error al enviar la presencia disponible

	// Caso: state != types.PresenceAvailable
	cli.sendActiveReceipts = 1
	err = cli.SendPresence(types.PresenceUnavailable)
	assert.NoError(t, err) // Verifica que no hay error al enviar la presencia no disponible
}

func TestSubscribePresence(t *testing.T) {
	cli := setupMockClient()

	// Caso: Error al obtener un token de privacidad
	cli.ErrorOnSubscribePresenceWithoutToken = true
	err := cli.SubscribePresence(types.JID("user@example.com"))
	assert.Error(t, err)

	// Caso: Suscripción exitosa con token de privacidad
	cli.ErrorOnSubscribePresenceWithoutToken = false
	cli.Store.PrivacyTokens = MockPrivacyTokenStore{}
	err = cli.SubscribePresence(types.JID("user@example.com"))
	assert.NoError(t, err)
}

func TestSendChatPresence(t *testing.T) {
	cli := setupMockClient()

	// Caso: getOwnID devuelve un ownID vacío
	cli.getOwnID = func() types.JID { return "" }
	err := cli.SendChatPresence(types.JID("user@example.com"), types.ChatPresenceComposing, "")
	assert.EqualError(t, err, ErrNotLoggedIn.Error())

	// Caso: sendNode se llama con los argumentos correctos
	cli.getOwnID = func() types.JID { return "12345@example.com" }
	err = cli.SendChatPresence(types.JID("user@example.com"), types.ChatPresenceComposing, "")
	assert.NoError(t, err) // Verifica que no hay error al enviar el estado de chat
}

func TestHandleChatState(t *testing.T) {
	cli := setupMockClient()

	// Caso: Nodo válido para una actualización de estado de chat
	node := &MockNode{}
	cli.handleChatState(node) // No hay errores esperados

	// Caso: Número inesperado de hijos en el nodo
	node = &MockNode{}
	node.On("GetChildren").Return([]waBinary.Node{{}, {}})
	cli.handleChatState(node) // Debería registrar una advertencia sobre el número inesperado de hijos

	// Caso: Estado de chat no reconocido
	node = &MockNode{}
	node.On("GetChildren").Return([]waBinary.Node{{Tag: "unknown_state"}})
	cli.handleChatState(node) // Debería registrar una advertencia sobre el estado de chat no reconocido
}

func TestHandlePresence(t *testing.T) {
	cli := setupMockClient()

	// Caso: Nodo válido para un evento de presencia
	node := &MockNode{}
	cli.handlePresence(node) // No hay errores esperados

	// Caso: Tipo de presencia no reconocido
	node = &MockNode{}
	node.On("AttrGetter").Return(&MockAttrGetter{})
	node.On("AttrGetter").Return(&MockAttrGetter{"type": "unknown_type"})
	cli.handlePresence(node) // Debería registrar una advertencia sobre el tipo de presencia no reconocido

	// Caso: Errores al analizar el evento de presencia
	node = &MockNode{}
	node.On("AttrGetter").Return(&MockAttrGetter{"last": "invalid_timestamp"})
	cli.handlePresence(node) // Debería registrar una advertencia sobre errores al analizar el evento de presencia
}
